cloud_function_src/jambase/requirements.txt
functions-framework==3.*
google-cloud-secret-manager==2.17.0
google-cloud-storage==2.14.0
requests==2.31.0
